requirements.txt
python-dotenv==1.0.1
psycopg[binary]==3.2.1
pydantic==2.8.2
pypdf==4.3.1
fastapi==0.115.0
uvicorn==0.30.6
google-generativeai==0.8.3
python-multipart==0.0.9
